`@remotion/jonnys-videos`: Separate video projects
