+++ packages/jonnys-videos/src/how-can-remotion-be-free/FirstCustomerChart.tsx
