+++ packages/jonnys-videos/src/how-can-remotion-be-free/HowCanRemotionBeFree.tsx
@@ -1,7 +1,10 @@
 import {Audio} from '@remotion/media';
 import React from 'react';
 import {Sequence, Solid} from 'remotion';
-import {AnimatedCaptions, CAPTIONS_HEIGHT} from './AnimatedCaptions';
+import {
+	AnimatedCaptions,
+	CAPTIONS_HEIGHT,
+} from '../components/AnimatedCaptions';
 import {asset} from './assets';
 import {
 	FIRST_CUSTOMER_CHART_DURATION_IN_FRAMES,
@@ -17,18 +20,20 @@ import {
 } from './UnrulyOutcome';
 import {ZURICH_PHOTO_DURATION_IN_FRAMES, ZurichPhoto} from './ZurichPhoto';
 
-const FREE_MASTER_AUDIO_FILE =
-	'ElevenLabs_2026-08-23T17_39_27_Derek - Fun & Energetic _pvc_sp115_s50_sb75_se0_b_m2.mp3';
-const FREE_MASTER_CAPTIONS_FILE = 'free-master-captions.json';
+const HOW_CAN_REMOTION_BE_FREE_AUDIO_FILE = 'voiceover.mp3';
+const HOW_CAN_REMOTION_BE_FREE_CAPTIONS_FILE = 'captions.json';
 const PREMIUM_VERSION_START_FRAME = 100;
 const UNRULY_OUTCOME_START_FRAME = 148;
 const FIRST_CUSTOMER_CHART_START_FRAME = 645;
 const SWITZERLAND_SCENE_START_FRAME = 816;
 
-export const FreeMaster: React.FC = () => {
+export const HowCanRemotionBeFree: React.FC = () => {
 	return (
 		<>
-			<Audio src={asset(FREE_MASTER_AUDIO_FILE)} durationInFrames={1674.71} />
+			<Audio
+				src={asset(HOW_CAN_REMOTION_BE_FREE_AUDIO_FILE)}
+				durationInFrames={1674.71}
+			/>
 			<Solid
 				width={1080}
 				height={1920}
@@ -79,8 +84,8 @@ export const FreeMaster: React.FC = () => {
 				}}
 			>
 				<AnimatedCaptions
-					captionsFile={FREE_MASTER_CAPTIONS_FILE}
-					voiceoverFile={null}
+					captionsSrc={asset(HOW_CAN_REMOTION_BE_FREE_CAPTIONS_FILE)}
+					voiceoverSrc={null}
 				/>
 			</Sequence>
 		</>
